secure the edit and delete product views

+++ products/views.py
@@ -3,6 +3,7 @@ from icecream import ic
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.db.models.functions import Lower
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 from products.models import Product, Category
 from products.forms import ProductForm
@@ -62,8 +63,13 @@ def product_detail(request, product_id):
     )
 
 
+@login_required
 def add_product(request):
     """add a product to the store"""
+
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that")
+        return redirect(reverse("home"))
 
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
@@ -85,8 +91,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """edit a product in the store"""
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that")
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.POST:
@@ -118,8 +129,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """delete a product form the store"""
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that")
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
